# Optimize ComputerInkBoundingBox(LtoR), remove unnecessary additional branch