## Description

Optimizes `ComputerInkBoundingBox` / `ComputerInkBoundingBoxLtoR` methods, collapsing additional branch for `TextFormattingMode.Display` into `EmGlyphMetrics`, which now takes a readonly `ref` of `GlyphMetrics` to avoid copies.

This is one of the hottest methods traditionally when working with glyphs, so it deserves a bit of special care. `EmGlyphMetrics` is just a data container with a bit of logic, and it is kind of imperative that the logic in the `ctor` is actually inlined inside the original method and the `struct` ceases to exist at runtime, otherwise due to the nature of the code, this could get pretty messy at the given JIT state.

This gets inlined usually even without forcing it via `AggressiveInlining` but it doesn't hurt to force it in this case, it would hurt in case it didn't get inlined, so we want to prevent that from happening.

For 8 glyphs, the difference is already significant:

| Method    | otherVal    | Mean [ns] | Error [ns] | StdDev [ns] | Code Size [B] | Allocated [B] |
|---------- |---------------- |----------:|-----------:|------------:|--------------:|--------------:|
| Original |  GlyphMetrics[8] |  42.41 ns |   0.251 ns |    0.222 ns |       2,882 B |             - |
| PR__EDIT  | GlyphMetrics[8] |  34.01 ns |   0.188 ns |    0.166 ns |       2,736 B |             - |

## Customer Impact

Improved performance.

## Regression

No.

## Testing

Local build, perf benchmarks.

## Risk

Imho it is low, we don't change any existing logic.